fix: ProductPostItem의 price 타입 수정 및 그에 따른 설정 파일 수정( String -> Int )

+++ app/src/main/java/com/mbj/ssassamarket/ui/writing/WritingViewModel.kt
@@ -119,7 +119,7 @@ class WritingViewModel @Inject constructor(private val postItemRepository: Produ
                     _productUploadResponse.value = postItemRepository.addProductPost(
                         title = title.value!!,
                         category = category.value!!,
-                        price = price.value!!,
+                        price = price.value!!.toInt(),
                         content = content.value!!,
                         imageLocations = selectedImageList.value.orEmpty(),
                         location = location.value!!,
